Refactor setup log utils
